fix: mark test as `skip` if before hooks failed
